fix: Restore "Open in Browser" button in ImageModal
- Add "openImageModal" function
- Add default image size option for "ViewIcons"
