+++ src/utils/discord.tsx
@@ -18,8 +18,10 @@
 
 import { MessageObject } from "@api/MessageEvents";
 import { findByPropsLazy, findLazy } from "@webpack";
-import { ChannelStore, ComponentDispatch, GuildStore, PrivateChannelsStore, SelectedChannelStore } from "@webpack/common";
+import { ChannelStore, ComponentDispatch, GuildStore, MaskedLink, ModalImageClasses, PrivateChannelsStore, SelectedChannelStore } from "@webpack/common";
 import { Guild, Message } from "discord-types/general";
+
+import { ImageModal, ModalRoot, ModalSize, openModalLazy } from "./modal";
 
 const PreloadedUserSettings = findLazy(m => m.ProtoClass?.typeName.endsWith("PreloadedUserSettings"));
 const MessageActions = findByPropsLazy("editMessage", "sendMessage");
@@ -76,4 +78,24 @@ export function sendMessage(
     };
 
     return MessageActions.sendMessage(channelId, messageData, waitForChannelReady, extra);
+}
+
+export function openImageModal(url: string, props?: Partial<React.ComponentProps<ImageModal>>): Promise<string> {
+    return openModalLazy(async () => _props => (
+        <ModalRoot
+            {..._props}
+            className={ModalImageClasses.modal}
+            size={ModalSize.DYNAMIC}>
+            <ImageModal
+                className={ModalImageClasses.image}
+                original={url}
+                placeholder={url}
+                src={url}
+                renderLinkComponent={props => <MaskedLink {...props} />}
+                shouldHideMediaOptions={false}
+                shouldAnimate
+                {...props}
+            />
+        </ModalRoot>
+    ));
 }
